add soft delete to videos

+++ app/Observers/VideoObserver.php
@@ -39,6 +39,8 @@ class VideoObserver
      */
     public function deleted(Video $video)
     {
+        if ($video->trashed()) return true;
+
         Storage::delete($video->path);
         Storage::delete($video->thumbnail);
     }
@@ -62,6 +64,7 @@ class VideoObserver
      */
     public function forceDeleted(Video $video)
     {
-        //
+        Storage::delete($video->path);
+        Storage::delete($video->thumbnail);
     }
 }
